# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Tarusho.Server",
        "description": "Tarusho server API reference",
        "version": "1.0.0"
    },
    "host": "localhost:21774",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "produces": [
        "application/json; charset=utf-8"
    ],
...
    "paths": {
        "/object_tags/{id}": {
            "get": {
                "summary": "指定したidのObjectTagに関する詳細を返します",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ObjectTagのId",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "ObjectTag"
                ],
                "responses": {
                    "200": {
                        "description": "指定したidのObjectTagに関する詳細",
                        "schema": {
                            "$ref": "#/definitions/ObjectTagResponse"
                        }
                    }
                }
            },
            "put": {
                "summary": "指定したidのObjectTagに関する詳細を変更します",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ObjectTagのId",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "item",
                        "in": "body",
                        "description": "変更を行う詳細情報(idの値に格納された要素を変更)",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ObjectTagRequest"
                        }
                    }
                ],
                "tags": [
                    "ObjectTag",
                    "Admin"
                ],
                "responses": {
                    "200": {
                        "description": "変更したObjectTagのオブジェクト",
                        "schema": {
                            "$ref": "#/definitions/ObjectTagResponse"
                        }
                    }
                }
            },
            "delete": {
                "summary": "指定したidのObjectTagに関する情報・予約を削除します",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ObjectTagのId",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "ObjectTag",
                    "Admin"
                ],
                "responses": {
                    "200": {
                        "description": "削除が完了した場合"
                    }
                }
            }
        },
...
        "/reservations/{id}": {
            "get": {
                "summary": "指定したidの予約に関する詳細を返します",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ReservationのId",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Reservation"
                ],
                "responses": {
                    "200": {
                        "description": "指定したidのReservationに関する詳細",
                        "schema": {
                            "$ref": "#/definitions/ReservationResponse"
                        }
                    }
                }
            },
            "put": {
                "summary": "指定したidのReservationに関する詳細を変更します",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ReservationのId",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "item",
                        "in": "body",
                        "description": "変更を行う詳細情報(idの値に格納された要素を変更)",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ReservationRequest"
                        }
                    }
                ],
                "tags": [
                    "Reservation",
                    "Admin"
                ],
                "responses": {
                    "200": {
                        "description": "変更したReservationのオブジェクト",
                        "schema": {
                            "$ref": "#/definitions/ReservationResponse"
                        }
                    }
                }
            },
            "delete": {
                "summary": "指定したidのReservationを削除します",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ReservationのId",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Reservation",
                    "Admin"
                ],
                "responses": {
                    "200": {
                        "description": "削除が完了した場合"
                    }
                }
            }
        },
...
        "/categories/{id}": {
            "get": {
                "summary": "指定したidのカテゴリに関する詳細を返します",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "CategoryのId",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "tags": [
                    "Category"
                ],
                "responses": {
                    "200": {
                        "description": "指定したidのCategoryに関する詳細",
                        "schema": {
                            "$ref": "#/definitions/CategoryRequest"
                        }
                    }
                }
            },
            "put": {
                "summary": "指定したidのCategoryに関する詳細を変更します",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "CategoryのId",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "item",
                        "in": "body",
                        "description": "変更を行う詳細情報(idの値に格納された要素を変更)",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CategoryResponse"
                        }
                    }
                ],
                "tags": [
                    "Category",
                    "Admin"
                ],
                "responses": {
                    "200": {
                        "description": "変更したCategoryのオブジェクト",
                        "schema": {
                            "$ref": "#/definitions/CategoryResponse"
                        }
                    }
                }
            },
            "delete": {
                "summary": "指定したidのCategory、および含まれる予約・ObjectTagを削除します",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "CategoryのId",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "tags": [
                    "Category",
                    "Admin"
                ],
                "responses": {
                    "200": {
                        "description": "削除が完了した場合"
                    }
                }
            }
        },
        "/categories": {
            "post": {
                "summary": "Categoryの新規登録を行います",
                "parameters": [
                    {
                        "name": "item",
                        "in": "body",
                        "description": "新規登録を行う際の詳細情報(ただしidの値は不要/サーバーが自動で採番)",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CategoryRequest"
                        }
                    }
                ],
                "tags": [
                    "Category",
                    "Admin"
                ],
                "responses": {
                    "201": {
                        "description": "作成したCategoryのオブジェクト",
                        "schema": {
                            "$ref": "#/definitions/CategoryResponse"
                        }
                    }
                }
            }
        },
...
        "/list/object_tags/{category_id}": {
            "get": {
                "summary": "Categoryに含まれるObjectTagの一覧を返します",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "description": "取得したいObjectTagが含まれるCategoryId",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "ObjectTag",
                    "PaginationEnabled"
                ],
                "responses": {
                    "200": {
                        "description": "検索結果のObjectTagのリスト",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ObjectTagResponse"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "IdNamePair": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "description": "Id field"
                },
                "name": {
                    "type": "string",
                    "description": "Name field"
                }
            }
        },
        "IdNumberNamePair": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "number",
                    "description": "Id field"
                },
                "name": {
                    "type": "string",
                    "description": "Name field"
                }
            }
        },
        "ObjectTagRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "description": "ObjectTagのId"
                },
                "name": {
                    "type": "string",
                    "description": "表示名"
                },
                "place": {
                    "type": "string",
                    "description": "場所"
                },
                "object_uri": {
                    "type": "string",
                    "description": "オブジェクトのURI"
                },
                "optional_uri": {
                    "type": "string",
                    "description": "Optional(バーコード)のURI"
                },
                "description": {
                    "type": "string",
                    "description": "説明文"
                },
                "category": {
                    "type": "number",
                    "description": "カテゴリId"
                },
                "booking_enabled": {
                    "type": "boolean",
                    "description": "将来の予約が可能か"
                }
            }
        },
        "ObjectTagResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "description": "ObjectTagのId"
                },
                "name": {
                    "type": "string",
                    "description": "表示名"
                },
                "place": {
                    "type": "string",
                    "description": "場所"
                },
                "object_uri": {
                    "type": "string",
                    "description": "オブジェクトのURI"
                },
                "optional_uri": {
                    "type": "string",
                    "description": "Optional(バーコード)のURI"
                },
                "description": {
                    "type": "string",
                    "description": "説明文"
                },
                "image_uri": {
                    "type": "string",
                    "description": "画像Uri"
                },
                "thumbnail_image_uri": {
                    "type": "string",
                    "description": "画像サムネイルuri"
                },
                "in_use_reservation_id": {
                    "type": "string",
                    "description": "現在利用中の予約Id"
                },
                "booking_enabled": {
                    "type": "boolean",
                    "description": "将来の予約が可能か"
                },
                "category": {
                    "$ref": "#/definitions/IdNumberNamePair"
                }
            }
        },
        "ReservationRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "description": "ReservationのId"
                },
                "comment": {
                    "type": "string",
                    "description": "予約時の要件"
                },
                "user": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "description": "使用者のユーザーId"
                },
                "object_tag_id": {
                    "type": "string",
                    "description": "予約するObjectTagのId"
                },
                "priority": {
                    "type": "number",
                    "description": "優先度(列挙型にする予定)"
                },
                "start_at": {
                    "type": "string",
                    "format": "date-time",
                    "description": "予約の開始日時"
                },
                "end_at": {
                    "type": "string",
                    "format": "date-time",
                    "description": "予約の終了日時"
                },
                "is_endless": {
                    "type": "boolean",
                    "description": "予約の終了が無期限であるかどうか"
                }
            }
        },
        "ReservationResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "description": "ReservationのId"
                },
                "comment": {
                    "type": "string",
                    "description": "Reservationのcomment"
                },
                "owner": {
                    "$ref": "#/definitions/IdNamePair"
                },
                "user": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/IdNamePair"
                    },
                    "description": "使用者のユーザーId, 表示名のリスト"
                },
                "object_tag": {
                    "$ref": "#/definitions/IdNamePair"
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time",
                    "description": "予約を行った日時"
                },
                "modified_at": {
                    "type": "string",
                    "format": "date-time",
                    "description": "最終更新日"
                },
                "priority": {
                    "type": "number",
                    "description": "優先度(列挙型にする予定)"
                },
                "start_at": {
                    "type": "string",
                    "format": "date-time",
                    "description": "予約の開始日時"
                },
                "end_at": {
                    "type": "string",
                    "format": "date-time",
                    "description": "予約の終了日時"
                },
                "is_endless": {
                    "type": "boolean",
                    "description": "予約の終了が無期限であるかどうか"
                }
            }
        },
        "CategoryRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64",
                    "description": "CategoryのId"
                },
                "description": {
                    "type": "string",
                    "description": "Categoryの説明"
                },
                "name": {
                    "type": "string",
                    "description": "Categoryの表示名"
                }
            }
        },
        "CategoryResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64",
                    "description": "CategoryのId"
                },
                "description": {
                    "type": "string",
                    "description": "Categoryの説明"
                },
                "name": {
                    "type": "string",
                    "description": "Categoryの表示名"
                }
            }
        },
...
    }
}